REPLACE subscriptionClient by operatorClient
In this commit:
- Refactored operatorClient
- Replaced subscription client
- Adjusted CRUD functions to receive namespace
- Moved Client stuff to client.go including interface
- Moved OperatorGroup data to operator_group.go
